requirements.txt
aliyun-python-sdk-core-v3==2.12.1
aliyun-python-sdk-kms==2.5.0
arrow==0.13.1
certifi==2018.11.29
chardet==3.0.4
crcmod==1.7
dataclasses==0.6
Django==1.11.11
django-redis==4.10.0
django-stubs==0.9.0
djangorestframework==3.9.0
djangorestframework-stubs==0.3.0
idna==2.8
isoweek==1.3.3
jmespath==0.9.3
mypy==0.670
mypy-extensions==0.4.1
numpy==1.15.4
oss2==2.6.0
pika==0.13.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycryptodome==3.7.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.7
redis==3.0.1
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
